requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-openai
langchain-core
openai
python-dotenv